Reworked locator
We have been having spurious crashes in this code. Although it appeared valid, maybe reworking it will fix them.
